Compelted Huawei Charging Support

+++ api/src/main/java/com/mobius/software/telco/protocols/diameter/primitives/creditcontrol/huawei/AccountInfo.java
@@ -79,7 +79,7 @@ public interface AccountInfo extends DiameterAvp
 	
 	void setAccountDate(ByteBuf value);
 	
-	Integer getLifeCycleState();
+	Long getLifeCycleState();
 	
-	void setLifeCycleState(Integer value);			
+	void setLifeCycleState(Long value);			
 }
